+++ demo_example/smtp/smtp_client.cpp
@@ -25,7 +25,7 @@
 #include <string_view>
 #include <vector>
 
-#include "../asio_util.hpp"
+#include "../asio_coro_util.hpp"
 
 #ifdef ENABLE_SSL
 #include "asio/ssl.hpp"
